[ch-163385912] add more APIs
