+++ app.js
@@ -12,13 +12,7 @@ import meetups from './routes/meetups';
 
 import comments from './routes/comments';
 
-// import upcomingMeetups from './routes/upcomingMeetups';
-//
-// import questions from './routes/questions';
-//
-// import upvote from './routes/upvoteQuestion';
-//
-// import downvote from './routes/downvoteQuestion';
+import questionsVotes from './routes/questionVotes';
 
 import users from './routes/users';
 
@@ -26,10 +20,7 @@ import home from './routes/home';
 
 const app = express();
 
-// get setting of our app
-
 app.use(bodyparser.json());
-
 
 app.use(express.json());
 
@@ -37,14 +28,10 @@ app.use(logger);
 app.use(helmet());
 
 app.use('/', home);
-// app.use('/api/v1/meetups', upcomingMeetups);
 app.use('/api/v1/meetups', meetups); // for any route started with /api/v1 use meetups router
 app.use('/api/v1/users', users);
 app.use('/api/v1/question/', comments);
-// app.use('/api/v1/meetups', questions);
-// app.use('/api/v1/questions', upvote);
-// app.use('/api/v1/questions', downvote);
-
+// app.use('/api/v1/question/', questionsVotes);
 
 if (app.get('env') === 'development') {
   app.use(morgan('tiny'));
